[JENKINS-20679] - Refactoring: Introduce a new AbstractJenkinsManifestMojo, move code there
